refactor: separate StoryOperations::ReadAll into its own file
