+++ app/operations/story_operations/read_all.rb
